+++ lib/services/storage_service.dart
@@ -12,6 +12,10 @@ abstract class StorageService {
 
   Future<Set<StorageVolume>> getStorageVolumes();
 
+  Future<Set<String>> getUntrackedTrashPaths(Iterable<String> knownPaths);
+
+  Future<Set<String>> getUntrackedVaultPaths(String vaultName, Iterable<String> knownPaths);
+
   Future<String> getVaultRoot();
 
   Future<int?> getFreeSpace(StorageVolume volume);
@@ -65,6 +69,33 @@ class PlatformStorageService implements StorageService {
     try {
       final result = await _platform.invokeMethod('getStorageVolumes');
       return (result as List).cast<Map>().map(StorageVolume.fromMap).toSet();
+    } on PlatformException catch (e, stack) {
+      await reportService.recordError(e, stack);
+    }
+    return {};
+  }
+
+  @override
+  Future<Set<String>> getUntrackedTrashPaths(Iterable<String> knownPaths) async {
+    try {
+      final result = await _platform.invokeMethod('getUntrackedTrashPaths', <String, dynamic>{
+        'knownPaths': knownPaths.toList(),
+      });
+      return (result as List).cast<String>().toSet();
+    } on PlatformException catch (e, stack) {
+      await reportService.recordError(e, stack);
+    }
+    return {};
+  }
+
+  @override
+  Future<Set<String>> getUntrackedVaultPaths(String vaultName, Iterable<String> knownPaths) async {
+    try {
+      final result = await _platform.invokeMethod('getUntrackedVaultPaths', <String, dynamic>{
+        'vault': vaultName,
+        'knownPaths': knownPaths.toList(),
+      });
+      return (result as List).cast<String>().toSet();
     } on PlatformException catch (e, stack) {
       await reportService.recordError(e, stack);
     }