+++ lib/widgets/about/app_ref.dart
@@ -1,4 +1,3 @@
-import 'dart:ui';
 
 import 'package:aves/model/device.dart';
 import 'package:aves/theme/icons.dart';