+++ lib/widgets/navigation/drawer/app_drawer.dart
@@ -1,4 +1,3 @@
-import 'dart:ui';
 
 import 'package:aves/model/filters/album.dart';
 import 'package:aves/model/filters/trash.dart';