+++ lib/widgets/settings/home_widget_settings_page.dart
@@ -1,4 +1,6 @@
+import 'package:aves/model/device.dart';
 import 'package:aves/model/filters/filters.dart';
+import 'package:aves/model/settings/enums/widget_outline.dart';
 import 'package:aves/model/settings/enums/widget_shape.dart';
 import 'package:aves/model/settings/settings.dart';
 import 'package:aves/services/widget_service.dart';
@@ -33,10 +35,11 @@ class HomeWidgetSettingsPage extends StatefulWidget {
 
 class _HomeWidgetSettingsPageState extends State<HomeWidgetSettingsPage> {
   late WidgetShape _shape;
-  late Color? _outline;
+  late WidgetOutline _outline;
   late WidgetOpenPage _openPage;
   late WidgetDisplayedItem _displayedItem;
   late Set<CollectionFilter> _collectionFilters;
+  Future<Map<Brightness, Map<WidgetOutline, Color?>>> _outlineColorsByBrightness = Future.value({});
 
   int get widgetId => widget.widgetId;
 
@@ -61,6 +64,24 @@ class _HomeWidgetSettingsPageState extends State<HomeWidgetSettingsPage> {
     _openPage = settings.getWidgetOpenPage(widgetId);
     _displayedItem = settings.getWidgetDisplayedItem(widgetId);
     _collectionFilters = settings.getWidgetCollectionFilters(widgetId);
+    WidgetsBinding.instance.addPostFrameCallback((_) => _updateOutlineColors());
+  }
+
+  void _updateOutlineColors() {
+    _outlineColorsByBrightness = _loadOutlineColors();
+    setState(() {});
+  }
+
+  Future<Map<Brightness, Map<WidgetOutline, Color?>>> _loadOutlineColors() async {
+    final byBrightness = <Brightness, Map<WidgetOutline, Color?>>{};
+    await Future.forEach(Brightness.values, (brightness) async {
+      final byOutline = <WidgetOutline, Color?>{};
+      await Future.forEach(WidgetOutline.values, (outline) async {
+        byOutline[outline] = await outline.color(brightness);
+      });
+      byBrightness[brightness] = byOutline;
+    });
+    return byBrightness;
   }
 
   @override
@@ -71,58 +92,70 @@ class _HomeWidgetSettingsPageState extends State<HomeWidgetSettingsPage> {
         title: Text(l10n.settingsWidgetPageTitle),
       ),
       body: SafeArea(
-        child: Column(
-          children: [
-            Expanded(
-              child: ListView(
-                children: [
-                  _buildShapeSelector(),
-                  ListTile(
-                    title: Text(l10n.settingsWidgetShowOutline),
-                    trailing: HomeWidgetOutlineSelector(
-                      getter: () => _outline,
-                      setter: (v) => setState(() => _outline = v),
-                    ),
+        child: FutureBuilder<Map<Brightness, Map<WidgetOutline, Color?>>>(
+          future: _outlineColorsByBrightness,
+          builder: (context, snapshot) {
+            final outlineColorsByBrightness = snapshot.data;
+            if (outlineColorsByBrightness == null) return const SizedBox();
+
+            final effectiveOutlineColors = outlineColorsByBrightness[Theme.of(context).brightness];
+            if (effectiveOutlineColors == null) return const SizedBox();
+
+            return Column(
+              children: [
+                Expanded(
+                  child: ListView(
+                    children: [
+                      _buildShapeSelector(effectiveOutlineColors),
+                      ListTile(
+                        title: Text(l10n.settingsWidgetShowOutline),
+                        trailing: HomeWidgetOutlineSelector(
+                          getter: () => _outline,
+                          setter: (v) => setState(() => _outline = v),
+                          outlineColorsByBrightness: outlineColorsByBrightness,
+                        ),
+                      ),
+                      SettingsSelectionListTile<WidgetOpenPage>(
+                        values: WidgetOpenPage.values,
+                        getName: (context, v) => v.getName(context),
+                        selector: (context, s) => _openPage,
+                        onSelection: (v) => setState(() => _openPage = v),
+                        tileTitle: l10n.settingsWidgetOpenPage,
+                      ),
+                      SettingsSelectionListTile<WidgetDisplayedItem>(
+                        values: WidgetDisplayedItem.values,
+                        getName: (context, v) => v.getName(context),
+                        selector: (context, s) => _displayedItem,
+                        onSelection: (v) => setState(() => _displayedItem = v),
+                        tileTitle: l10n.settingsWidgetDisplayedItem,
+                      ),
+                      SettingsCollectionTile(
+                        filters: _collectionFilters,
+                        onSelection: (v) => setState(() => _collectionFilters = v),
+                      ),
+                    ],
                   ),
-                  SettingsSelectionListTile<WidgetOpenPage>(
-                    values: WidgetOpenPage.values,
-                    getName: (context, v) => v.getName(context),
-                    selector: (context, s) => _openPage,
-                    onSelection: (v) => setState(() => _openPage = v),
-                    tileTitle: l10n.settingsWidgetOpenPage,
+                ),
+                const Divider(height: 0),
+                Padding(
+                  padding: const EdgeInsets.all(8),
+                  child: AvesOutlinedButton(
+                    label: l10n.saveTooltip,
+                    onPressed: () {
+                      _saveSettings();
+                      WidgetService.configure();
+                    },
                   ),
-                  SettingsSelectionListTile<WidgetDisplayedItem>(
-                    values: WidgetDisplayedItem.values,
-                    getName: (context, v) => v.getName(context),
-                    selector: (context, s) => _displayedItem,
-                    onSelection: (v) => setState(() => _displayedItem = v),
-                    tileTitle: l10n.settingsWidgetDisplayedItem,
-                  ),
-                  SettingsCollectionTile(
-                    filters: _collectionFilters,
-                    onSelection: (v) => setState(() => _collectionFilters = v),
-                  ),
-                ],
-              ),
-            ),
-            const Divider(height: 0),
-            Padding(
-              padding: const EdgeInsets.all(8),
-              child: AvesOutlinedButton(
-                label: l10n.saveTooltip,
-                onPressed: () {
-                  _saveSettings();
-                  WidgetService.configure();
-                },
-              ),
-            ),
-          ],
+                ),
+              ],
+            );
+          },
         ),
       ),
     );
   }
 
-  Widget _buildShapeSelector() {
+  Widget _buildShapeSelector(Map<WidgetOutline, Color?> outlineColors) {
     return Padding(
       padding: const EdgeInsets.symmetric(vertical: 8, horizontal: 8),
       child: Wrap(
@@ -143,7 +176,7 @@ class _HomeWidgetSettingsPageState extends State<HomeWidgetSettingsPage> {
                 height: 124,
                 decoration: ShapeDecoration(
                   gradient: selected ? gradient : deselectedGradient,
-                  shape: _WidgetShapeBorder(_outline, shape),
+                  shape: _WidgetShapeBorder(_outline, shape, outlineColors),
                 ),
               ),
             ),
@@ -169,12 +202,13 @@ class _HomeWidgetSettingsPageState extends State<HomeWidgetSettingsPage> {
 }
 
 class _WidgetShapeBorder extends ShapeBorder {
-  final Color? outline;
+  final WidgetOutline outline;
   final WidgetShape shape;
+  final Map<WidgetOutline, Color?> outlineColors;
 
   static const _devicePixelRatio = 1.0;
 
-  const _WidgetShapeBorder(this.outline, this.shape);
+  const _WidgetShapeBorder(this.outline, this.shape, this.outlineColors);
 
   @override
   EdgeInsetsGeometry get dimensions => EdgeInsets.zero;
@@ -191,10 +225,11 @@ class _WidgetShapeBorder extends ShapeBorder {
 
   @override
   void paint(Canvas canvas, Rect rect, {TextDirection? textDirection}) {
-    if (outline != null) {
+    final outlineColor = outlineColors[outline];
+    if (outlineColor != null) {
       final path = shape.path(rect.size, _devicePixelRatio);
       canvas.clipPath(path);
-      HomeWidgetPainter.drawOutline(canvas, path, _devicePixelRatio, outline!);
+      HomeWidgetPainter.drawOutline(canvas, path, _devicePixelRatio, outlineColor);
     }
   }
 
@@ -203,13 +238,15 @@ class _WidgetShapeBorder extends ShapeBorder {
 }
 
 class HomeWidgetOutlineSelector extends StatefulWidget {
-  final ValueGetter<Color?> getter;
-  final ValueSetter<Color?> setter;
+  final ValueGetter<WidgetOutline> getter;
+  final ValueSetter<WidgetOutline> setter;
+  final Map<Brightness, Map<WidgetOutline, Color?>> outlineColorsByBrightness;
 
   const HomeWidgetOutlineSelector({
     super.key,
     required this.getter,
     required this.setter,
+    required this.outlineColorsByBrightness,
   });
 
   @override
@@ -217,35 +254,40 @@ class HomeWidgetOutlineSelector extends StatefulWidget {
 }
 
 class _HomeWidgetOutlineSelectorState extends State<HomeWidgetOutlineSelector> {
-  static const List<Color?> options = [
-    null,
-    Colors.black,
-    Colors.white,
-  ];
-
   @override
   Widget build(BuildContext context) {
     return DropdownButtonHideUnderline(
-      child: DropdownButton<Color?>(
+      child: DropdownButton<WidgetOutline>(
         items: _buildItems(context),
         value: widget.getter(),
         onChanged: (selected) {
-          widget.setter(selected);
+          widget.setter(selected ?? WidgetOutline.none);
           setState(() {});
         },
       ),
     );
   }
 
-  List<DropdownMenuItem<Color?>> _buildItems(BuildContext context) {
-    return options.map((selected) {
-      return DropdownMenuItem<Color?>(
+  List<DropdownMenuItem<WidgetOutline>> _buildItems(BuildContext context) {
+    return supportedWidgetOutlines.map((selected) {
+      final lightColors = widget.outlineColorsByBrightness[Brightness.light];
+      final darkColors = widget.outlineColorsByBrightness[Brightness.dark];
+      return DropdownMenuItem<WidgetOutline>(
         value: selected,
         child: ColorIndicator(
-          value: selected,
-          child: selected == null ? const Icon(AIcons.clear) : null,
+          value: lightColors?[selected],
+          alternate: darkColors?[selected],
+          child: lightColors?[selected] == null ? const Icon(AIcons.clear) : null,
         ),
       );
     }).toList();
   }
+
+  List<WidgetOutline> get supportedWidgetOutlines => [
+        WidgetOutline.none,
+        WidgetOutline.black,
+        WidgetOutline.white,
+        WidgetOutline.systemBlackAndWhite,
+        if (device.isDynamicColorAvailable) WidgetOutline.systemDynamic,
+      ];
 }